Gerrit script - less hackish hack to get the room list
Look for any HUBOT_<whatever>_ROOMS environment variable to
determine the room list instead of hardcoding just a few
variable names. Still a hack, but it should work for all the
currently listed adapters that take a list of room and probably
will work for any new adapters added later.