[Janky] Replace all spaces in user's name when requesting a new build.
Janky is returning a 400 (Bad Request) error at times when a request for a new
build is made. Those errors are due to whitespaces that are being used in user's
names for campfire and possibly another chat engines.
Replacement of all spaces for plus signs "+" in user's names makes the URL that
is being posted to from hubot's janky script compliant to the specification [1]
and solves the issue.

[1] http://www.w3.org/Addressing/URL/url-spec.txt