Updates jenkins.coffee to respond correctly to latest jenkins update and adds convenience method to build.
Jenkins changed the http status code for their REST API in the last version. These changes handle responses better when requesting a build. It also adds a convenient way to start build, by keeping an internal array of jobs, so that the user can request a build by sending a number instead of a job's name.