Use :? instead of ( is)? in the listeners
While I prefer the sentence structure using the word is, the
roles.coffee script heavily listens for and responds to commands with
is in them. This causes a very confusing message when updating cash:

    > hubot cash burn rate is $10,000
    Ok, our burn rate is $10,000 per month
    I don't know anything about cash burn rate

Since roles.coffee is shipped with hubot and used in a number of
other situations, I've concluded that the best course of action is to
change the listeners for this script.

I've explored other options for this, such as using "set burn rate at"
or "update burn rate to", but this seems like the simplest change
with the least amount of impact.