Fix regular expression in plusplus script to properly match umlauts
The previous regular expression would not correctly match inputs like "Björn ++" or "Björn --", because it expected a group of word-characters or whitespace-characters followed by a group of non-word-characters or non-whitespace-characters.

With this fix the regular expression now expects a group of word-characters or non-space-characters followed by a group of non-word-characters or space-characters.