github-commit-link: Add support for repo@<SHA> and user/repo@<SHA>
Based on similar parsing in github-issue-link.coffee.  The main change
in the logic is that the initial .* is now the non-greedy .*?, so it
doesn't gobble up the repo-part of the regexp [1].  The user/repo
portion is [^\s@]+, which matches one or more characters that are
neither whitespace [2] nor @.

The non-capturing (?:...) block [3] around the optional repo@ or
user/repo@ qualifier allows us to group the qualifier with its @, even
though we only need the repo or user/repo (without the @) in our
internal logic.

I use CoffeeScript's block regular expressions to comment the regexp
[4], and CoffeeScript's existential operator '?' to check if
'msg.match[1]' is defined [5].

I also document the fact that lines matching 'commit/' are ignored.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions#special-questionmark
[2]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions#special-white-space
[3]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions#special-non-capturing-parentheses
[4]: http://coffeescript.org/#regexes
[5]: http://coffeescript.org/#operators