github-commit-link: Only show the first line of the commit
This is traditionally the message summary.  From Git's own docs [1]:

> The first line of the commit message should be a short description
> (50 characters is the soft limit, see DISCUSSION in git-commit(1)),
> and should skip the full stop.

I sometimes write long, detailed commit messages (like this one), and
there's no need to spam the channel with all of this motivational
background.  The summary line should be sufficient context there, and
folks looking for more information can follow the URL (which I've
moved to it's own line).

[1]: https://github.com/git/git/blob/v1.9.0/Documentation/SubmittingPatches#L89