Update jira-issues.coffee
This change fixes a problem where unassigned tickets would result in "[*ERROR*] TypeError: Cannot read property '0' of undefined". Our copy of this file referred to the possibility of asignee properties being kept in json.fields.assignee.values rather than json.fields.assignee. Not sure if that was a custom changed that never ended up in a merge request or just a remnant of an older version. This change should be compatible with both possibilities either way.