redis-brain: Check if info.path is null before replacing strings
If the optional prefix path is not given, redis-brain currently dies
with:

  ERROR Unable to load /.../redis-brain: TypeError: Cannot call method 'replace' of null
    at module.exports (/.../redis-brain.coffee:28:12, <js>:13:24)
    at Robot.loadFile (/.../robot.coffee:216:9, <js>:162:24)
    at Robot.loadHubotScripts (/.../robot.coffee:243:7, <js>:195:28)
    at /.../hubot:98:15, <js>:99:30
    at fs.js:266:14
    at Object.oncomplete (fs.js:107:15)

because:

  Url.parse "redis://localhost:6379"

returns a null path:

  { protocol: 'redis:',
    slashes: true,
    auth: null,
    host: 'localhost:6379',
    port: '6379',
    hostname: 'localhost',
    hash: null,
    search: null,
    query: null,
    pathname: null,
    path: null,
    href: 'redis://localhost:6379' }

Avoid that by only calling info.path.replace if info.path is truthy.

Neither this implementation nor the one it replaces will let you
explicitly avoid setting a prefix, but that's probably not a big deal.